[Behat] Extract coupon transformer
